% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correttor-package.R
\docType{package}
\name{correttor-package}
\alias{correttor}
\alias{correttor-package}
\title{correttor: Corretto Java Installer for R User}
\description{
Provides install function of corretto, one of the Java distributions.
             Package scope is Windows & MacOS.
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
